Fixes: #31 move helpers to a new file
